chippy: use libxkbcommon for keyboard input; fix memory corruption
We used to convert xcb key events into X11 ones so that we could use
Xutf8LookupString and XLookupString, respectively, and even though
older versions of QT and SFML did this it apperently can no longer be
done as it causes memory corruption and tricky crashes.

Now we use libxkbcommon (statically linked because many linux distros
do not provide the correct version yet) to manage keyboard state and
convert keycodes into keysyms and then into unicode input.

The downside is that libxkbcommon does not operate as an IME, nor does
it support composed character input. This means further work is to be
done to properly support e.g. Asian langauges etc.