chippy: pass *Display as void* due to incompatible headers on systems
It seems different GLX headers define the Display type differently
which causes compiler errors on some systems. Since we define all
GLX procedures ourselves (so we can support EGL in the future) we
can just change the data type to void* and be careful passing the
type around as usual.

Fixes issue 57.