gfx/gl2: add keepState renderer option for cooperation with other renderers
Some renderers may require that the state is exactly the way it was before
some external OpenGL code was ran, or might be buggy otherwise. This adds
a boolean option 'keepState' to gl2.New which specifies if the OpenGL state
should be maintained between frames.

Essentially, this means each frame the existing OpenGL state is queried,
our frame is rendered, and the old OpenGL state is restored. This will help
ensure no issues with e.g. QT5 occur where lingering state on our part
causes QT5 to render incorrectly.